modify: modify the way you use context API
